requirements.txt
keras==2.7.0
matplotlib==3.6.2
numpy==1.24.0
opencv_python==4.6.0.66
pandas==1.4.4
Pillow==9.3.0
streamlit==1.16.0
streamlit_drawable_canvas==0.9.2
tensorflow==2.7.0
